1. CConfigurable object flimits use floats not double. 2. CNADC2530 was attempting to init -ipl, -vector with 0 not"0"

+++ llnlReadout/devices/CConfigurableObject.h
@@ -118,7 +118,7 @@ public:
 
   struct  flimit {
     bool   s_checkMe;
-    double s_value;
+    float s_value;
     flimit() :
       s_checkMe(false) {}
     flimit(float value) :
